requirements.txt
streamlit>=0.84.0
Pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.11.0
scikit-learn>=0.24.2